Example support for multiple flows
